Fix "depositing started again when ending pet summoning"
An S_VIEW_WARE_EX packet with action attibute set to 1 (closing window) 
is received when ending pet summoning. And therefore started depositing 
before. Now, only action set to 0 is allowed for depositing.
Note: If there are other actions for depositing then this must be 
changed again to "action is not 1"